added a tab layout for the trending fragment with a viewpager under populated with the former trending feed fragment, worked on styling, fixed trending book title capitalization, worked on overall styling
