+++ app/src/main/res/layout/fragment_home.xml
@@ -25,5 +25,4 @@
         android:layout_height="match_parent"
         android:id="@+id/viewPager"
         android:layout_below="@+id/tabLayout"/>
-
 </RelativeLayout>
