+++ app/src/main/res/layout/item_review.xml
@@ -12,7 +12,8 @@
         style="?attr/materialCardViewFilledStyle"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:clickable="true">
+        android:clickable="true"
+        app:cardBackgroundColor="@color/primary_red_lighter">
 
         <Button
             android:id="@+id/bAddToList"
@@ -27,49 +28,30 @@
             android:id="@+id/tvUsername"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="75dp"
-            android:layout_marginTop="4dp"
+            android:layout_marginStart="84dp"
+            android:layout_marginTop="25dp"
+            android:textSize="20sp"
             app:layout_constraintStart_toEndOf="@+id/ivProfilePic"
             app:layout_constraintTop_toTopOf="parent"
             tools:text="ReallyCoolUsername" />
 
         <com.google.android.material.imageview.ShapeableImageView
             android:id="@+id/ivProfilePic"
-            android:layout_width="56dp"
-            android:layout_height="56dp"
-            android:layout_marginStart="4dp"
-            android:layout_marginTop="4dp"
+            android:layout_width="65dp"
+            android:layout_height="65dp"
+            android:layout_marginStart="10dp"
+            android:layout_marginTop="10dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:shapeAppearanceOverlay="@style/ShapeAppearanceOverlay.App.circleImageView"
             tools:srcCompat="@tools:sample/avatars" />
 
         <TextView
-            android:id="@+id/tvReviewTitle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="75dp"
-            android:layout_marginTop="30dp"
-            app:layout_constraintStart_toEndOf="@+id/ivProfilePic"
-            app:layout_constraintTop_toBottomOf="@+id/tvUsername"
-            tools:text="A must watch comedy!" />
-
-        <TextView
-            android:id="@+id/tvReviewBody"
-            android:layout_width="201dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="75dp"
-            android:layout_marginTop="60dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tvMediaType"
-            tools:text="review body......" />
-
-        <TextView
             android:id="@+id/tvMediaTitle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="75dp"
-            android:layout_marginTop="120dp"
+            android:layout_marginTop="80dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/tvReviewTitle"
             tools:text="Oh Hello on Broadway" />
@@ -79,17 +61,17 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="75dp"
-            android:layout_marginTop="145dp"
+            android:layout_marginTop="100dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/tvMediaTitle"
             tools:text="Movie" />
 
         <ImageView
             android:id="@+id/ivMediaPic"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
-            android:layout_marginStart="30dp"
-            android:layout_marginTop="180dp"
+            android:layout_width="350dp"
+            android:layout_height="350dp"
+            android:layout_marginStart="15dp"
+            android:layout_marginTop="125dp"
             app:srcCompat="@drawable/poster_placeholder"
             tools:srcCompat="@drawable/poster_placeholder" />
 
@@ -115,8 +97,8 @@
             android:id="@+id/ibAddToList"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="285dp"
-            android:layout_marginTop="435dp"
+            android:layout_marginLeft="310dp"
+            android:layout_marginTop="425dp"
             android:background="@android:color/transparent"
             android:src="@drawable/add_icon_material"
             tools:src="@drawable/add_icon_material"
