+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="primary_green">#6C7E5B</color>
-    <color name="primary_green_variant">#4D5545</color>
+    <color name="primary_red">#EF5350</color>
+    <color name="primary_red_darker">#E53935</color>
+    <color name="primary_red_lighter">#FFEBEE</color>
     <color name="secondary_orange">#D57023</color>
     <color name="secondary_orange_variant">#A1561E</color>
     <color name="purple_200">#FFBB86FC</color>
