+++ app/src/main/res/values/themes.xml
@@ -2,19 +2,25 @@
     <!-- Base application theme. -->
     <style name="Theme.Captstone" parent="Theme.Material3.Light">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/primary_green</item>
-        <item name="colorPrimaryVariant">@color/primary_green_variant</item>
+        <item name="colorPrimary">@color/primary_red</item>
+        <item name="colorPrimaryVariant">@color/primary_red_darker</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/secondary_orange</item>
         <item name="colorSecondaryVariant">@color/secondary_orange_variant</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
+        <item name="android:statusBarColor" tools:targetApi="l">@color/primary_red</item>
         <!-- Customize your theme here. -->
+        <item name="colorSurface">@color/white</item>
     </style>
 
     <style name="ShapeAppearanceOverlay.App.circleImageView" parent="">
         <item name="cornerSize">50%</item>
     </style>
+
+    <style name="App.Custom.Indicator" parent="Widget.Material3.BottomNavigationView.ActiveIndicator">
+        <item name="android:color">@color/primary_red_lighter</item>
+    </style>
+
 </resources>
